Move fulltext event listeners
This move is made just to avoid circular imports when trying in
initialize the fulltext event listeners.

This commit just moves the code and fixes the import paths for the new
file to work
